Move mute-user-list-item component
