+++ apps/web/src/components/domains/user/mute-user-list-item/mute-user-list-item.css.ts
@@ -1,6 +1,6 @@
 import { style } from '@vanilla-extract/css';
 
-import { theme } from '../../../../../themes';
+import { theme } from '../../../../themes';
 
 export const wrapper = style({
   display: 'flex',
